+++ R/write.R
@@ -220,11 +220,30 @@ write_data2 <- function(
   dc <- openxlsx2_type(data)
 
   # convert factor to character
-  if (any(dc == openxlsx2_celltype[["factor"]])) {
-    is_factor <- dc == openxlsx2_celltype[["factor"]]
+  is_factor <- dc == openxlsx2_celltype[["factor"]]
+
+  if (any(is_factor)) {
     fcts <- names(dc[is_factor])
     data[fcts] <- lapply(data[fcts], to_string)
-    # dc <- openxlsx2_type(data)
+  }
+
+  # remove xml encoding and reapply it afterwards. until v0.3 encoding was not enforced.
+  # until 1.1 formula encoding was applied in write_formula() and missed formulas written
+  # as data frames with class formula
+  is_fml <- dc %in% c(
+    openxlsx2_celltype[["formula"]], openxlsx2_celltype[["array_formula"]],
+    openxlsx2_celltype[["cm_formula"]], openxlsx2_celltype[["hyperlink"]]
+  )
+
+  if (any(is_fml)) {
+    fmls <- names(dc[is_fml])
+    data[fmls] <- lapply(
+      data[fmls],
+      function(val) {
+        val <- replaceXMLEntities(val)
+        vapply(val, function(x) xml_value(xml_node_create("fml", x, escapes = TRUE), "fml"), "")
+      }
+    )
   }
 
   hconvert_date1904 <- grepl('date1904="1"|date1904="true"',
@@ -467,9 +486,9 @@ write_data2 <- function(
 
     # options("openxlsx2.numFmt" = NULL)
     if (any(dc == openxlsx2_celltype[["numeric"]])) { # numeric or integer
-      if (!is.null(unlist(options("openxlsx2.numFmt")))) {
+      if (!is.null(getOption("openxlsx2.numFmt"))) {
 
-        numfmt_numeric <- unlist(options("openxlsx2.numFmt"))
+        numfmt_numeric <- getOption("openxlsx2.numFmt")
 
         dim_sel <- get_data_class_dims("numeric")
         # message("numeric: ", dim_sel)
@@ -482,11 +501,7 @@ write_data2 <- function(
       }
     }
     if (any(dc == openxlsx2_celltype[["short_date"]])) { # Date
-      if (is.null(unlist(options("openxlsx2.dateFormat")))) {
-        numfmt_dt <- 14
-      } else {
-        numfmt_dt <- unlist(options("openxlsx2.dateFormat"))
-      }
+      numfmt_dt <- getOption("openxlsx2.dateFormat", 14)
 
       dim_sel <- get_data_class_dims("short_date")
       # message("short_date: ", dim_sel)
@@ -498,11 +513,7 @@ write_data2 <- function(
       )
     }
     if (any(dc == openxlsx2_celltype[["long_date"]])) {
-      if (is.null(unlist(options("openxlsx2.datetimeFormat")))) {
-        numfmt_posix <- 22
-      } else {
-        numfmt_posix <- unlist(options("openxlsx2.datetimeFormat"))
-      }
+      numfmt_posix <- getOption("openxlsx2.datetimeFormat", default = 22)
 
       dim_sel <- get_data_class_dims("long_date")
       # message("long_date: ", dim_sel)
@@ -514,11 +525,7 @@ write_data2 <- function(
       )
     }
     if (any(dc == openxlsx2_celltype[["hms_time"]])) {
-      if (is.null(unlist(options("openxlsx2.hmsFormat")))) {
-        numfmt_hms <- 21
-      } else {
-        numfmt_hms <- unlist(options("openxlsx2.hmsFormat"))
-      }
+      numfmt_hms <- getOption("openxlsx2.hmsFormat", default = 21)
 
       dim_sel <- get_data_class_dims("hms_time")
       # message("hms: ", dim_sel)
@@ -530,11 +537,7 @@ write_data2 <- function(
       )
     }
     if (any(dc == openxlsx2_celltype[["accounting"]])) { # accounting
-      if (is.null(unlist(options("openxlsx2.accountingFormat")))) {
-        numfmt_accounting <- 4
-      } else {
-        numfmt_accounting <- unlist(options("openxlsx2.accountingFormat"))
-      }
+      numfmt_accounting <- getOption("openxlsx2.accountingFormat", default = 4)
 
       dim_sel <- get_data_class_dims("accounting")
       # message("accounting: ", dim_sel)
@@ -545,11 +548,7 @@ write_data2 <- function(
       )
     }
     if (any(dc == openxlsx2_celltype[["percentage"]])) { # percentage
-      if (is.null(unlist(options("openxlsx2.percentageFormat")))) {
-        numfmt_percentage <- 10
-      } else {
-        numfmt_percentage <- unlist(options("openxlsx2.percentageFormat"))
-      }
+      numfmt_percentage <- getOption("openxlsx2.percentageFormat", default = 10)
 
       dim_sel <- get_data_class_dims("percentage")
       # message("percentage: ", dim_sel)
@@ -561,11 +560,7 @@ write_data2 <- function(
       )
     }
     if (any(dc == openxlsx2_celltype[["scientific"]])) {
-      if (is.null(unlist(options("openxlsx2.scientificFormat")))) {
-        numfmt_scientific <- 48
-      } else {
-        numfmt_scientific <- unlist(options("openxlsx2.scientificFormat"))
-      }
+      numfmt_scientific <- getOption("openxlsx2.scientificFormat", default = 48)
 
       dim_sel <- get_data_class_dims("scientific")
       # message("scientific: ", dim_sel)
@@ -577,11 +572,7 @@ write_data2 <- function(
       )
     }
     if (any(dc == openxlsx2_celltype[["comma"]])) {
-      if (is.null(unlist(options("openxlsx2.comma")))) {
-        numfmt_comma <- 3
-      } else {
-        numfmt_comma <- unlist(options("openxlsx2.commaFormat"))
-      }
+      numfmt_comma <- getOption("openxlsx2.commaFormat", default = 3)
 
       dim_sel <- get_data_class_dims("comma")
       # message("comma: ", dim_sel)
@@ -1032,9 +1023,6 @@ write_formula <- function(
   standardize_case_names(...)
 
   assert_class(x, "character")
-  # remove xml encoding and reapply it afterwards. until v0.3 encoding was not enforced
-  x <- replaceXMLEntities(x)
-  x <- vapply(x, function(val) xml_value(xml_node_create("fml", val, escapes = TRUE), "fml"), NA_character_)
   dfx <- data.frame("X" = x, stringsAsFactors = FALSE)
 
   formula <- "formula"