+++ NEWS.md
@@ -1,3 +1,22 @@
+# openxlsx2 1.2
+
+## New features
+
+* Add new params to `wb_add_pivot_table()`. It is now possible to set the `show_data_as` value and set a tabular table design. [833](https://github.com/JanMarvin/openxlsx2/pull/833)
+
+## Fixes
+
+* Previously formulas written as data frames were not xml escaped. [834](https://github.com/JanMarvin/openxlsx2/pull/834)
+
+* Improve drawing relationship id selection that could cause issues with unordered relationship ids in loaded workbooks. [838](https://github.com/JanMarvin/openxlsx2/pull/838)
+
+* Improve copying cells in transpose mode and with hyperlinks. [850](https://github.com/JanMarvin/openxlsx2/pull/850)
+
+* Options `openxlsx2.maxWidth` and `openxlsx2.minWidth` are now respected as documented when setting column widths with `wb_set_col_widths()`. [847](https://github.com/JanMarvin/openxlsx2/issues/847)
+
+***************************************************************************
+
+
 # openxlsx2 1.1
 
 ## New features
@@ -50,7 +69,7 @@ They will continue to work for some time, but changing to newer functions is rec
 
 * `wb_comment()` is a new helper function to help create `wbComment` objects
   It is very similar to `create_comment()`, with the following differences:
-  * `author` looks at `options("openxlsx2.creator")`; (`create_comment()` only used `sys.getenv("user")`)
+  * `author` looks at `options("openxlsx2.creator")`; (`create_comment()` only used `sys.getenv("USERNAME")` / `Sys.getenv("USER")`)
   * `visible` defaults to `FALSE` to account for modern spreadsheet software behavior. (`create_comment()`, it is `TRUE`).
   * `width` and `height` must now be of length 1. (In `create_comment()`, the first element is taken, other are ignored.)